Implement the `ManifestManager`.
The `ManifestManager` creates the `ManifestHeader` and the manifest 'body',
which is stored in an SQLite database.

The `ManifestHeader` contains the index of the next outstanding job, the order
in which the jobs are scheduled to run, and the total number of reboots
performed during the experiment, while the body stores the status of each job.